requirements.txt
graphics.py==5.0.1.post1
pandas>=2.0,<3.0
openpyxl>=3.1,<4.0
matplotlib>=3.7,<4.0